add: x_close, x_open
